Refactor HighlightViewModal to smaller components
