+++ frontend/src/components/BookPage/components/HighlightViewModal/components/HighlightContent.tsx
@@ -1,0 +1,77 @@
+import type { Highlight } from '@/api/generated/model';
+import { CalendarMonth as CalendarIcon, FormatQuote as QuoteIcon } from '@mui/icons-material';
+import { Box, Typography } from '@mui/material';
+
+interface HighlightContentProps {
+  highlight: Highlight;
+}
+
+export const HighlightContent = ({ highlight }: HighlightContentProps) => {
+  const startsWithLowercase =
+    highlight.text.length > 0 &&
+    highlight.text[0] === highlight.text[0].toLowerCase() &&
+    highlight.text[0] !== highlight.text[0].toUpperCase();
+
+  const renderHighlightText = () => {
+    const prefix = startsWithLowercase ? '...' : '';
+    // Split by newlines and filter out empty strings
+    const paragraphs = highlight.text.split('\n').filter((p) => p.trim() !== '');
+
+    return (
+      <>
+        {paragraphs.map((paragraph, index) => (
+          <Typography
+            key={index}
+            component="p"
+            variant="h6"
+            sx={{
+              fontWeight: 500,
+              color: 'text.primary',
+              lineHeight: 1.7,
+              margin: 0,
+              marginBottom: index < paragraphs.length - 1 ? 2 : 0,
+            }}
+          >
+            {index === 0 ? prefix + paragraph : paragraph}
+          </Typography>
+        ))}
+      </>
+    );
+  };
+
+  return (
+    <Box display="flex" flexDirection="column" gap={3} flex={1}>
+      {/* Highlight Text */}
+      <Box sx={{ display: 'flex', alignItems: 'start', gap: 2 }}>
+        <QuoteIcon
+          sx={{
+            fontSize: 28,
+            color: 'primary.main',
+            flexShrink: 0,
+            mt: 0.5,
+            opacity: 0.7,
+          }}
+        />
+        <Box sx={{ flex: 1 }}>{renderHighlightText()}</Box>
+      </Box>
+
+      {/* Metadata */}
+      <Box sx={{ display: 'flex', gap: 2, alignItems: 'center', opacity: 0.8 }}>
+        <CalendarIcon
+          sx={{
+            fontSize: 20,
+            color: 'text.secondary',
+          }}
+        />
+        <Typography variant="body2" color="text.secondary">
+          {new Date(highlight.datetime).toLocaleDateString('en-US', {
+            year: 'numeric',
+            month: 'long',
+            day: 'numeric',
+          })}
+          {highlight.page && ` • Page ${highlight.page}`}
+        </Typography>
+      </Box>
+    </Box>
+  );
+};
