+++ frontend/src/components/BookPage/components/HighlightViewModal/index.ts
@@ -1,0 +1,2 @@
+export { HighlightViewModal } from './HighlightViewModal';
+export type { HighlightViewModalProps } from './HighlightViewModal';
